refactor: extract types to separate module
